DCC++ESP32 BASE STATION, initial revision.
All basic functionality has been tested and working except for the web
interface for the programmer which will be addressed in an upcoming
commit if necessary. The web inteface throttle needs a bit more work as
it has a few known issues in it, such as duplicate listener for the
websocket data when toggling the console connect button from the Base
Station tab vs Throttle tab.