Updates to the web interface for turnouts, removal of SD support.
Creating/Updating a turnout using only the DCC address of the decoder with "DCC Address Only" selected will now correctly calculate the board address / index automatically. This simplifies the approach for addressing to be inline with other base stations.

Removing the SD support for configuration at this point, it is not used/tested, SPIFFS has been extensively tested and is working sufficiently for now.