start of LCC integration, not complete so do not enable.
OpenMRN library is required for LCC and it is not currently
included (yet).

Added the ability to lookup a turnout by its dcc address. Note this
is not the same as <T ID ADDR SUB-ADDR>, it will not take SUB-ADDR into
consideration. It should be used for cases where the turnout has been
created using DCC address only, ie: <T ID ADDR -1> (or via the web
interface)