Major update
1) restructured libraries (again), NeoNextion and OpenMRN now compile
cleanly from the lib tree as they should. LocoNet2 still TBD.
2) programming track will no longer power up unless it is actively being
used.
3) web interface cleanup to bring it inline with previous versions where
editing a sensor/turnout/output/etc will bring up the existing details.
4) LCC is now available for use with a handful of events being
processed, there are limitations on this today which I'm still working
on removing. Namely, if you have track power ON you can not access the
configuration dialog within JMRI without the command station crashing
and restarting.
5) DCC signal generation has been adjusted slightly to make moving to a
new model for generating the signal easier (hopefully). The current
usage of an ISR is a problem as it accesses (and modifies) tracking
details that reside outside of IRAM/DRAM which is a problem when SPIFFS
(spi_flash) is accessed.
6) programming track now follows RCN-216 guidelines for packets to send
and timing for them (not exact but getting closer).
7) programming track is now rate limited at 300mA which is a 20% buffer
over the RCN-215 guideline of 250mA.