Fixes for turnout orientation and wifi adjustments.
Turnout orientation was being corrupted when converted to a json payload
for the web interface and for persistence. This appears to be a bug in
the way that enums are converted by ArduinoJson. Instead of relying on
automatic conversion it is now casted as an int so it will always be
converted correctly.

Turnout display on web interface has been updated to now consolidate
address/subaddress into one column and turnout orientation is now
displayed in the turnout list instead.

WiFi connection was timing out occasionally due to the call to
WiFi.waitForConnectResult() returning after 10sec without a successful
connection or DHCP IP address assignment. Since this process can take
longer on some routers a while loop is now in place to allow more time
for the connection process to complete, this usually takes under 20sec.