fix #7 and drop experimental inverted signal pin support.
Adding the BTS7960B connections and standalone signal split circuit used
for both the BTS7960B and Pololu MC33926 Motor Driver Carrier boards.
The experimental support for inverted signal pin on the ESP32 is not
supportable long term due to the post-v1.2.0 planned redesign of the DCC
signal generation code. If the redesign will support the inverted pin it
will be re-added at that time.