Adding OTA support to the web interface.
This will allow uploading a new firmware without needing to directly
connect to the command station. The web interface will need a few more
updates but it is functional. The primary issue with the web interface
is that it will redirect to a broken link while the command station is
rebooting.