Nextion 3.5" display support and log level adjustments
Adding Nextion 3.5" display support in addition to the 3.2" display
support. The screen is identified on startup of the command station with
a default size of 3.2". The primary difference between the two screen
sizes is the number of turnouts that can be displayed (15 vs 24).

Lowered log level for most messages to reduce console output as it is
not generally enabled/used.