Fix NextionInterface issue related to invalidating a locomotive.
When a locomotive was invalidated via the web interface it calls the
NextionInterface (if enabled) to remove the locomotive from the
throttle. Unfortunately the NextionInterface code was using -1 to denote
an invalid locomotive and this value was being stored in a uint16_t
which resulted in the address being stored as 65535 instead of the
expected "no loco". Added guard code to ensure this doesn't occur now.