background task removal, wifi scan added.
removing background tasks for HC12 and LCC as there are a few too many
tasks running which is leading to delays due to task switching, more
updates are coming for this.

Added a WiFi scan to display the available SSIDs on the serial console
when the WiFi connection is not successful. This is the first step
towards SoftAP for configuration portal.