Updating Nextion interface...
The 5.0" display support is now detected correctly and turnout count
reflected on the turnout page when 5.0" display is detected.

Updated the wifi startup code to remain on the title screen with wifi
status being updated on the screen until the IP has been provisioned,
this way the user doesn't attempt to use the nextion until the CS is
fully ready for use. This will also give status updates every ~10sec
during the wifi connect process.