Fixing bug with S88 via web interface, switching log_X to LOG
S88 sensor bus creation was failing due to invalid json payload as well
as the section always being hidden regardless of the CS having the
feature enabled.

Switching from log_x to LOG (and LOG_ERROR) to clean up the console
diagnostic output. This is still a WIP but is mostly done now.