Adjusting the motor board code to enable/disable the DCC signal itself
If you enable a single motor board output via the web interface it would
not enable the DCC signal leading to an unusable DCC signal. The only
way to enable the signal was via enable-all/disable-all (sending <1> or
<0> via JMRI or clicking on the power toggle in web interface). Now if
you click on the power icon for an individual motor board it will
enable/disable the DCC signal.