Allow reconfiguration of the DCC preamble bits
The defaults are 16 (OPS) and 22 (PROG). In previous releases both OPS
and PROG would always use 22 which is more than required for OPS and is
the minimum for PROG. Some older sound decoders may require more bits on
the PROG track which via configuration can increase this up to 50.

OPS is restricted between 11 and 20, PROG is restricted between 22 and
50.