replacing tab with spaces, adjusting RMT task priority
There was a mix of tab and spaces across multiple source files, standardizing on spaces. Still to do is cleanup two spaces vs four for indents.

Increased RMT task priority and pinned to core zero, same as OpenMRN. It is still to be tested if this will cause conflicts for concurrent execution.