:ambulance: Replace space with + in the WS Key.
In some cases the WS client may not be using the base64 URL safe character set. When this is the case a "+" will be replaced with " " (space) as part of url_decode processing. For the WS key used for handshake this needs to be retained otherwise the client will close the connection immediately following handshake due to decode failure.

Fixes https://github.com/atanisoft/ESP32CommandStation/issues/48