:sparkles: Add ability to reset the LCC uplink from the web interface
The last known uplink address is persisted in a hidden field when reconnect is enabled. This works generally well for most cases but in some cases it results in non-desirable behavior with no clear exit strategy. This commit adds the ability to clear the cached address but does not address an underlying problem with the reconnect strategy where it does not give up trying to reconnect and fallback to mDNS based search after N attempts.

https://github.com/atanisoft/ESP32CommandStation/issues/69