:tractor: Switch to FreeRTOS Queue for RMT
The usage of Atomic to protect the DeviceBuffer can lead to unnecessary delays when the lock is taken inside an ISR context. This is due to the SMP nature of the esp32 since it stops interrupts and tasks on both cores.

Since the lock blocks both cores it causes the TWAI driver to start dropping frames when more than ~80/sec are received. With this change the performance is returned to ~1800/sec.