Prioritized DCC Packet support (#85)
* :tractor: Introduce prioritized packet scheduler.

* :pencil: send e-stop packets until cleared

This changes the original behavior to be toggled rather than send a burst of e-stop packets and stop.

* :pencil: Track enabled state for e-stop

* :pencil: rename shutdown method

* :pencil: Explicitly call notify() rather than use AutoNotify

* :pencil: Initialize exclusive index member var

* :pencil: Fixes for priority updates support

* :pencil: Remove usage of OSMutex in train DB

* :tractor: Switch to FreeRTOS Queue for RMT

The usage of Atomic to protect the DeviceBuffer can lead to unnecessary delays when the lock is taken inside an ISR context. This is due to the SMP nature of the esp32 since it stops interrupts and tasks on both cores.

Since the lock blocks both cores it causes the TWAI driver to start dropping frames when more than ~80/sec are received. With this change the performance is returned to ~1800/sec.

* :pencil: Enable FreeRTOS static allocation support

* :pencil: portYIELD_FROM_ISR with arg is IDF v4.3+