:pencil: Adjust the DCC packet encoding.
The RMT end-of-packet marker bit was not being accounted for in the pktLength_ variable which resulted in the RMT not knowing it reached the end of the packet. This bug was introduced as part of the EMC spectrum spreading support which has now been reworked to be done after fully encoding the packet.