:sparkles: Adding coredump capture/conversion.
During startup the CS will verify if a coredump is present in flash, if one is found the CS will set LEDs to yellow/red alterating blinking pattern for up to 15sec during which time if the factory reset button is held for at least one second the coredump will be erased. If the factory reset button is not held the CS will halt with the LEDs remaining in a blinking pattern.